refactor: projectModal 파일 styled 코드 분리 및 props 코드 리팩토링
